Petty Moe (Male-2-Female Natural Voice)
Petty Moe Female Android (Male-2-Female Natural Voice) is a copyright-free merge of several high-quality female voices. It performs particularly well with a European accents. This free AI voice can be utilized freely with tools like https://github.com/w-okada/voice-changer or similar platforms. The voice is published for educational use and not for malicious purposes. The author disclaims responsibility for any misuse of this AI voice, which is provided for research purposes.